Added fuzziness to metal. Metal, Lambertian and Material class. Added Albedo (diffuse/attenuation color).

+++ Amarelo/src/Math/Interval.h
@@ -7,20 +7,20 @@ class Interval
 {
 public:
 	Interval() : m_Min(+Amrl::g_AmrlInfinity), m_Max(-Amrl::g_AmrlInfinity) {};
-	Interval(double min, double max) : m_Min(min), m_Max(max) {};
+	Interval(float min, float max) : m_Min(min), m_Max(max) {};
 
-	inline double Size() const { return (m_Max- m_Min); }
+	inline float Size() const { return (m_Max- m_Min); }
 
-	inline bool Contains(double x) const { return (m_Min <= x && x <= m_Max); }
-	inline bool Surrounds(double x) const { return (m_Min < x && x < m_Max); }
+	inline bool Contains(float x) const { return (m_Min <= x && x <= m_Max); }
+	inline bool Surrounds(float x) const { return (m_Min < x && x < m_Max); }
 
-	double Clamp(double x) const { return Amrl::clamp<double>(m_Min, m_Max, x); }
+	float Clamp(float x) const { return Amrl::clamp<float>(m_Min, m_Max, x); }
 
-	inline const double Max() const { return m_Max; }
-	inline const double Min() const { return m_Min; }
+	inline const float Max() const { return m_Max; }
+	inline const float Min() const { return m_Min; }
 
 private:
-	double m_Min, m_Max;
+	float m_Min, m_Max;
 };
 
 extern const Interval g_Empty;
